+++ springjdbcResultSet/src/main/java/com/spring/jdbc/EmployeeRowMapper.java
@@ -12,7 +12,7 @@ public class EmployeeRowMapper implements RowMapper<Employee> {
 		Employee e = new Employee();
 		e.setId(rs.getInt(1));
 		e.setName(rs.getString(2));
-		e.setSalary(rs.getDouble(3));
+		e.setSalary(rs.getInt(3));
 		return e;
 	}
